ai-service/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-anthropic
langchain-community
python-dotenv
pydantic
